requirements .txt
# Core Dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-google-genai>=0.0.6
langchain-chroma>=0.1.0
google-generativeai>=0.3.0
chromadb>=0.4.0

# Document Processing
pypdf
python-docx
unstructured

# Utilities
python-dotenv
tqdm

# Optional: For better PDF processing
pdfplumber

